backend/requirements.txt
# FastAPI и веб-сервер
fastapi==0.104.1
uvicorn[standard]==0.24.0

# База данных и ORM
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
greenlet==3.0.1
alembic==1.12.1
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.27.0

# Безопасность
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# Утилиты
python-dotenv==1.0.0
APScheduler==3.10.1
python-json-logger==2.0.7

# Тестирование
pytest==7.4.3
httpx==0.25.2

# CORS
fastapi-cors==0.0.6 